requirements.txt
streamlit
plotly
gitpython
radon
pandas
numpy
scikit-learn